backend/requirements.txt
fastapi
pymongo
uvicorn
python-dotenv
openai
fastmcp
# openai-agents